requirements.txt
apex==0.9.10dev
cuml==0.6.1.post1
decord==0.6.0
deepspeed==0.16.3
ftfy==6.3.1
fvcore==0.1.5.post20221221
imageio==2.34.2
lmdb==1.3.0
matplotlib==3.8.4
mmcv==2.2.0
mmdet==3.3.0
mmseg==1.3.0
muon_optimizer==0.1.0
netlib==0.17
numpy==2.3.5
omegaconf==2.3.0
opencv_python==4.9.0.80
packaging==25.0
pandas==2.3.3
Pillow==12.0.0
regex==2024.11.6
scikit_learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
setuptools==69.5.1
skimage==0.0
submitit==1.5.3
tensorboardX==2.6.2.2
tensorboardX==2.6.4
tensorflow==2.20.0
torch==2.2.0
torchmetrics==1.1.2
torchvision==0.17.0
tqdm==4.66.2
wandb==0.23.0
xformers==0.0.24
yacs==0.1.8
